pyproject.toml
[project]
name = "aurora-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-import-export>=4.3.7",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-standardized-errors>=0.14.1",
    "openpyxl>=3.1.5",
    "pytelegrambotapi>=4.26.0",
]

[dependency-groups]
dev = [
    "django-silk>=5.3.2",
    "factory-boy>=3.3.3",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "aurora_backend.settings"
